fix: address review feedback for #57
