+++ src/client/routes/Todos/utils/dateUtils.ts
@@ -99,3 +99,42 @@ export function getQuickDates() {
         nextWeek: nextWeek.toISOString(),
     };
 }
+
+/**
+ * Format timestamp as relative time with absolute fallback
+ * @param dateString - ISO date string to format
+ * @returns Formatted relative time string (e.g., "2 hours ago", "Yesterday at 3:45 PM", "Jan 25 at 3:45 PM")
+ */
+export function formatRelativeTime(dateString: string): string {
+    const date = new Date(dateString);
+    const now = new Date();
+    const diffMs = now.getTime() - date.getTime();
+    const diffMins = Math.floor(diffMs / 60000);
+    const diffHours = Math.floor(diffMs / 3600000);
+    const diffDays = Math.floor(diffMs / 86400000);
+
+    const formatAbsolute = (d: Date) => {
+        const thisYear = now.getFullYear();
+        const dateYear = d.getFullYear();
+        const month = d.toLocaleDateString('en-US', { month: 'short' });
+        const day = d.getDate();
+        const time = d.toLocaleTimeString('en-US', { hour: 'numeric', minute: '2-digit', hour12: true });
+
+        if (dateYear === thisYear) {
+            return `${month} ${day} at ${time}`;
+        }
+        return `${month} ${day}, ${dateYear} at ${time}`;
+    };
+
+    if (diffMins < 1) {
+        return 'Just now';
+    } else if (diffMins < 60) {
+        return `${diffMins} ${diffMins === 1 ? 'minute' : 'minutes'} ago`;
+    } else if (diffHours < 24) {
+        return `${diffHours} ${diffHours === 1 ? 'hour' : 'hours'} ago`;
+    } else if (diffDays === 1) {
+        return `Yesterday at ${date.toLocaleTimeString('en-US', { hour: 'numeric', minute: '2-digit', hour12: true })}`;
+    } else {
+        return formatAbsolute(date);
+    }
+}
